+++ database/migrations/2020_10_19_112601_create_lesson_order_table.php
@@ -17,7 +17,7 @@ class CreateLessonOrderTable extends Migration
             $table->bigIncrements('id');
             $table->integer('quantity');
             $table->float('price');
-            $table->unsignedBigInteger('lesson_id');
+            $table->unsignedInteger('lesson_id');
             $table->unsignedBigInteger('order_id');
             $table->timestamps();
